% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{check_prob}
\alias{check_prob}
\title{Check Prob}
\usage{
check_prob(prob)
}
\arguments{
\item{prob}{vector of probabilities for each side}
}
\value{
TRUE if vector is a numeric vector with more than 1 element, 0<x<1, and elements add to 1. Throws error otherwise.
}
\description{
checks the validity of argument prob in device function
}
