% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{check_sides}
\alias{check_sides}
\title{Check Sides}
\usage{
check_sides(sides)
}
\arguments{
\item{sides}{vector of k ≥ 2 elements}
}
\value{
TRUE if sides is a vector with more than 1 element, not containing duplicates. Throws error otherwise.
}
\description{
checks the validity of argument sides in device function
}
