% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{check_times}
\alias{check_times}
\title{Check Times}
\usage{
check_times(times)
}
\arguments{
\item{times}{integer}
}
\value{
TRUE if times is positive integer greater than or equal to 1, ERROR otherwise
}
\description{
checks the validity of argument times in device function
}
