% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{data_count}
\alias{data_count}
\title{data_count}
\usage{
data_count(x)
}
\arguments{
\item{x}{object}
}
\value{
list of counting rolling object \code{"data_count"}
}
\description{
counting of rolling object \code{"data_count"}
}
