% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{device}
\alias{device}
\title{Device}
\usage{
device(sides = c(1, 2), prob = c(0.5, 0.5))
}
\arguments{
\item{sides}{vector of k ≥ 2 elements, by default numbers 1 and 2}

\item{prob}{vector of probabilities for each side (all equal to 1/2 by default).}
}
\value{
device object
}
\description{
creates a coin with sides and probability, coin is fair by default
}
