% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{is.device}
\alias{is.device}
\title{Is Device}
\usage{
is.device(x)
}
\arguments{
\item{x}{an object}
}
\value{
TRUE if object x is of class 'device', FALSE otherwise
}
\description{
checks if the object is of class device
}
