% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{is.rolls}
\alias{is.rolls}
\title{is.rolls function}
\usage{
is.rolls(rolls)
}
\arguments{
\item{rolls}{any input}
}
\value{
TRUE if the input is of class "roll" or stops/gives an error otherwise
}
\description{
checks if the input is of the class "rolls" or not
}
