% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{make_roll}
\alias{make_roll}
\title{Make Roll Object}
\usage{
make_roll(device, throw)
}
\arguments{
\item{device}{object of class device}

\item{throw}{object of class throw}
}
\description{
Constructor function for object "roll"
}
\keyword{internal}
