% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalmethods.R
\name{make_rolls}
\alias{make_rolls}
\title{make_rolls}
\usage{
make_rolls(roll, rolls)
}
\arguments{
\item{roll}{a list of roll information}

\item{rolls}{roll outcomes}
}
\value{
amended side, count, probability summary of the input roll
}
\description{
amend changes in parameters of rolls
}
