% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-rolls.R
\name{plot.roll}
\alias{plot.roll}
\title{Plot of object rolls}
\usage{
\method{plot}{roll}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"device"}}

\item{\dots}{arguments to be passed to/from other methods}
}
\description{
Plots the relative frequencies of a series of rolls
}
\examples{

 # create a device and roll it 40 times
 device1 <- device()
 roll40 <- roll(device1, times = 40)
 
 plot(roll40)

}
