% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{plot.rolls}
\alias{plot.rolls}
\title{Plot Rolls}
\usage{
\method{plot}{rolls}(x)
}
\arguments{
\item{x}{a 'rolls' object}
}
\value{
the barchart of the proportions
}
\description{
graphs a barchart of the proportions of the sides of devices
}
