% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{+.rolls}
\alias{+.rolls}
\title{Roll Addition Method}
\usage{
\method{+}{rolls}(x, i)
}
\arguments{
\item{x}{the 'roll' object}

\item{i}{the number of rolls to be added}
}
\value{
adds i number of rolls to the roll object x
}
\description{
adds more rolls
}
