% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{print.device}
\alias{print.device}
\title{Print Device}
\usage{
\method{print}{device}(x)
}
\arguments{
\item{x}{a device object}
}
\value{
prints the data frame of the device object with its sides and probabilities
}
\description{
prints the device object
}
