% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{print.rolls}
\alias{print.rolls}
\title{Print Rolls}
\usage{
\method{print}{rolls}(x)
}
\arguments{
\item{x}{a 'roll' object}
}
\value{
the name of the object, as well as the generated rolls
}
\description{
prints the 'rolls' object
}
