% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{print.summary.rolls}
\alias{print.summary.rolls}
\title{Print Summary Rolls}
\usage{
\method{print}{summary.rolls}(x)
}
\arguments{
\item{x}{a 'summary.roll' object}
}
\value{
the name of the object, as well as the dataframe of side, count, and prop associated with it
}
\description{
prints the summary.rolls object's frequency dataframe
}
