% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{roll}
\alias{roll}
\title{Rolls}
\usage{
roll(device, times = 1L)
}
\arguments{
\item{device}{object of class device (a vector)}

\item{times}{number of tosses (default value of 1)}
}
\value{
an object of class "rolls"
}
\description{
creates and rolls the device for a number of times
}
