% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{[.rolls}
\alias{[.rolls}
\title{Roll Extraction Method}
\usage{
\method{[}{rolls}(x, i)
}
\arguments{
\item{x}{the 'roll' object}

\item{i}{the index}
}
\value{
the index of the roll
}
\description{
extracts the index of the roll
}
