% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-roll-methods.R
\name{[<-.roll}
\alias{[<-.roll}
\title{[<-}
\usage{
\method{[}{roll}(x, i) <- value
}
\arguments{
\item{x}{an object of class \code{"roll"}}

\item{i}{an integer}

\item{value}{an element in sides}
}
\description{
replace the value of a given roll.
}
