% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{[<-.rolls}
\alias{[<-.rolls}
\title{Roll Replacement Method}
\usage{
\method{[}{rolls}(x, i) <- value
}
\arguments{
\item{x}{the 'roll' object}

\item{i}{the index}

\item{value}{the value to replace the index}
}
\value{
the new value at the index
}
\description{
replaces the value at index
}
