% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{summary.rolls}
\alias{summary.rolls}
\title{Summary Rolls}
\usage{
\method{summary}{rolls}(x)
}
\arguments{
\item{x}{a 'roll' object}
}
\value{
a dataframe with sides (of rolled 'device'),
       counts (frequency of the number of the roll), and prop (the relative frequencies of each side of the rolled 'device')
}
\description{
takes a 'roll' object and makes a summary of all the data associated with it
}
