% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{check_prob}
\alias{check_prob}
\title{check_prob}
\usage{
check_prob(prob)
}
\arguments{
\item{prob}{probability of sides}
}
\value{
TRUE/FALSE if the probabilities are legit or not
}
\description{
check if probability of the object is between 0 and 1, summed up to be 1
}
