% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{check_prob}
\alias{check_prob}
\title{check prob}
\usage{
check_prob(prob)
}
\arguments{
\item{probability}{for each side of the die (numeric vector)}
}
\value{
TRUE or FALSE
}
\description{
checks if probability input is valid
}
