% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{check_prob}
\alias{check_prob}
\title{Check_Prob}
\usage{
check_prob(prob)
}
\arguments{
\item{prob}{numerical vector}
}
\value{
error message if not valid
}
\description{
aux function: checks validity of 'prob' argument in
device()
}
