% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{check_prob}
\alias{check_prob}
\title{check_prob function}
\usage{
check_prob(prob)
}
\arguments{
\item{prob}{vector of probabilities for each side}
}
\value{
TRUE if conditions are met or stops with error message if any one of the conditions are not met
}
\description{
checks if the probability has any NA values, each probability is between 0 and 1 inclusive, and the elements in the probability add up to 1
}
\examples{
# check_prob(c(.5, .4, .1))
# [1] TRUE

# check_prob(c(.2, .2, .2, .2, .2))
# [1] TRUE

# check_prob(c(-1, .4, .1))
# Error in check_prob(c(-1, 0.4, 0.1)) :
# prob' values must range within 0 and 1

}
