% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{check_prob}
\alias{check_prob}
\title{Check_prob}
\usage{
check_prob(prob)
}
\arguments{
\item{prob}{vector of prob}
}
\value{
TRUE if prob is right prob \code{"check_prob"}
}
\description{
check for right prob \code{"check_prob"}
}
\examples{
check_prob(0.5)

}
