% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{check_sides}
\alias{check_sides}
\title{check_sides}
\usage{
check_sides(sides)
}
\arguments{
\item{sides}{vector of object sides}
}
\value{
TRUE/FALSE if the sides are legit or not
}
\description{
check if sides of the object are non-duplicated positive number
}
