% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{check_sides}
\alias{check_sides}
\title{check sides}
\usage{
check_sides(x)
}
\arguments{
\item{x}{vector with each side element}
}
\value{
returns boolean value of whether given sides are valid side values
}
\description{
checks the validity of the argument sides. sides must be a vecotr with more than one element with no duplicates
}
