% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{check_sides}
\alias{check_sides}
\title{Check_Sides}
\usage{
check_sides(sides)
}
\arguments{
\item{sides}{any vector that has length greater than 1}
}
\value{
error message if not valid
}
\description{
aux function: checks validity of 'sides' argument in device()
}
