% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{check_sides}
\alias{check_sides}
\title{Check_sides}
\usage{
check_sides(sides)
}
\arguments{
\item{sides}{vector of sides}
}
\value{
TRUE if sides is right sides \code{"check_sides"}
}
\description{
check for right sides \code{"check_sides"}
}
\examples{
sides = c('i', 'ii', 'iii', 'iv')
check_sides(sides)
}
