% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{check_times}
\alias{check_times}
\title{check_times}
\usage{
check_times(times)
}
\arguments{
\item{times}{input time}
}
\value{
TRUE/FALSE of if time is qualified
}
\description{
check if time is positive integer larger or equal to 1
}
