% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{check_times}
\alias{check_times}
\title{check times}
\usage{
check_times(times)
}
\arguments{
\item{number}{of times rolled (int)}
}
\value{
TRUE or FALSE
}
\description{
checks to see if the die is rolled a valid number of times
}
