% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{check_times}
\alias{check_times}
\title{check_times function}
\usage{
check_times(times)
}
\arguments{
\item{times}{a numeric input}
}
\value{
TRUE if times is a positive integer or FALSE is times is not a positive integer
}
\description{
checks if the number of times is a positive integer
}
\examples{
# check_times(4)
# [1] TRUE

# check_times(100)
# [1] TRUE

}
