% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{check_times}
\alias{check_times}
\title{check times}
\usage{
check_times(times)
}
\arguments{
\item{times}{number of rolling}
}
\value{
checking right times. If it is TRUE right times \code{"check_times"}
}
\description{
checking right times \code{"check_times"}
}
\examples{
check_times(2)


}
