% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{data_count}
\alias{data_count}
\title{data_count}
\usage{
data_count(x)
}
\arguments{
\item{rolling}{object}
}
\value{
summary of count, prop, and sides \code{"data_count"}
}
\description{
summary of rolling object and get data frame of sides, prob,
rolls, and total \code{"data_count"}
}
\examples{
data_count()

}
