% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{device}
\alias{device}
\title{device}
\usage{
device(sides = c(1, 2), prob = c(0.5, 0.5))
}
\arguments{
\item{sides}{vector of sides}

\item{prob}{vector of side probabilities}
}
\value{
an object of class
}
\description{
Creates an object of class
}
