% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{device}
\alias{device}
\alias{is.device}
\title{device}
\usage{
device(sides = c(1, 2), prob = c(0.5, 0.5))

is.device(x)
}
\arguments{
\item{sides}{vector of coin sides (k >= 2 elements)}

\item{prob}{vector of side probabilities}

\item{x}{an R object}
}
\value{
an object of class \code{"device"}
}
\description{
Creates an object of class \code{"device"}
}
\examples{
# default
device1 <- device()

# another
weird_die <- device(sides = c("i", "ii", "iii", "iv"),
prob = rep(1/4,4))

# loaded die
loaded_die <- device(sides = 1:6, rep = c(0.075,0.1,0.125,0.15,0.2,0.35))

}
\seealso{
\code{\link{roll}}
}
