% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{device}
\alias{device}
\title{device}
\usage{
device(sides = c(1, 2), prob = c(0.5, 0.5))
}
\arguments{
\item{sides}{sides info of the object}

\item{prob}{prob info of the object}
}
\value{
device information, including sides and probabilities
}
\description{
return information of device
}
