% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{device}
\alias{device}
\title{device constructor}
\usage{
device(sides = c(1, 2), prob = c(0.5, 0.5))
}
\arguments{
\item{number}{of sides (numeric vector)}

\item{probability}{of each side (numeric vector)}
}
\value{
a device object
}
\description{
creates an object called 'device'
}
