% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{device}
\alias{device}
\title{device}
\usage{
device(sides = c(1, 2), prob = rep(1/2, length(sides)))
}
\arguments{
\item{sides}{vector of k>=2 elements, by default numbers 1 and 2}

\item{prob}{vector of probabilities for each side (all equal to 1/2 by default)}
}
\value{
creates a coin with given sides and probabilities
}
\description{
creates a coin device
}
\examples{
#default
dev <- device()

#typical die
dice <- device(sides = c(1,2,3,4,5,6), prob = rep(1/6, 6))

#unfair coin
coin <- device(sides = c('head', 'tail'), prob = c(0.3, 0.7))
}
