% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roller.R
\name{device}
\alias{device}
\title{Device}
\usage{
device(sides = 1:2, prob = c(0.5, 0.5))
}
\arguments{
\item{sides}{a character or numeric vector that will act as the names of the sides of the device}

\item{prob}{a numeric vector of decimals that sum to one that will act as the probabilities of the sides of the device}
}
\value{
an object of class \code{"device"}
}
\description{
Creates an object of class \code{"device"}
}
\examples{

default_device <- device()

loded_die <- device(sides = 1:6, prob = c(.1, .1, .1, .1, .1, .5))

fair_die <- device(sides = 1:6, prob = rep(1/6, 6))

named_coin <- device(sides = c("heads", "tails", "left", "right"), prob = rep(1/4, 4))

}
