% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{device}
\alias{device}
\alias{is.device}
\title{Device}
\usage{
device(sides = c(1, 2), prob = c(0.5, 0.5))

is.device(x)
}
\arguments{
\item{sides}{a vector containing the side numbers of the object, which are always positive integers with 1 being the minimum}

\item{prob}{a vector containing the corresponding probabilities of the sides of the object}

\item{x}{an object in R}
}
\value{
a "device" object
}
\description{
returns a "device" object which has a certain number of sides and a distribution of probabilities for landing on those sides
}
\examples{
#default
device1 <- device(sides = c(1,2), prob = c(0.5,0.5))

#another example
device2 <- device(side = c(1,2,3), prob = c(0.1,0.3,0.6))
}
