% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{device}
\alias{device}
\title{Device Function}
\usage{
device(sides = c(1, 2), prob = rep(0.5, 2))
}
\arguments{
\item{sides}{any vector}

\item{prob}{numeric vector}
}
\value{
data frame indicating sides and corresponding
probabilities
}
\description{
Creates an object of class 'device'
}
\examples{

#fair die
device(sides = 1:6,
       prob = rep(1/6, 6)

#loaded die
device(sides = 1:6,
       prob = c(.15, .2, .3, .15, .1, .1))

#weird die
device(sides = c(letters[1:3], 2:5),
       prob = rep(1/7,7))
}
