% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{device}
\alias{device}
\alias{is.device}
\title{Device}
\usage{
device(sides = c(1, 2), prob = c(0.5, 0.5))

is.device(x)
}
\arguments{
\item{sides}{vector of device sides}

\item{prob}{vector of side probabilities}

\item{x}{an R object}
}
\value{
an object of class \code{"device"}
}
\description{
Creates an object of class \code{"device"}
}
\examples{
# default 
device1 <- device()

# another device
device <- device(c('heads', 'tails'))

# loaded device
loaded <- device(prob = c(0.7, 0.3))

}
