% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{device}
\alias{device}
\title{device.R}
\usage{
device(sides = c(1, 2), prob = c(0.5, 0.5))
}
\arguments{
\item{sides}{vector naming the sides of object}

\item{prob}{numeric vector of probabilities for each side}
}
\value{
an object of class \code{"device"}
}
\description{
create an object "device"
}
