% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{device}
\alias{device}
\alias{is.device}
\title{device}
\usage{
device(sides = c(1, 2), prob = c(0.5, 0.5))

is.device(x)
}
\arguments{
\item{sides}{vector of device}

\item{prob}{vector of probabilities}

\item{x}{an R object
#' @export}
}
\value{
an object of class device
}
\description{
Creates an object of class device
}
