% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{device}
\alias{device}
\alias{is.device}
\title{device}
\usage{
device(sides = c(1, 2), prob = c(1/2, 1/2))

is.device(x)
}
\arguments{
\item{sides}{vector of k ≥ 2 elements, by default numbers 1 and 2.}

\item{prob}{vector of probabilities for each side (all equal to 1/2 by default), must be the same length as sides.}

\item{x}{an R object}
}
\value{
an object of class \code{"device"}
}
\description{
Creates an object of class \code{"device"}
}
