% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{device}
\alias{device}
\title{Device}
\usage{
device(sides = c(1, 2), prob = c(0.5, 0.5))
}
\arguments{
\item{sides}{vector of k >= 2 elements by default numbers 1 and 2, which must be of the same length as prob}

\item{prob}{vector of probabilities for each side (all equal to 1/2 by default), which must be of the same length as sides}
}
\value{
displays the class of the object, and a tabular display of the sides and the associated probabilities
}
\description{
stimulates rollin and object using the object of class "device"
}
\examples{
# default
# fair_coin <- device()

# another device
# regular die
# regular_die <- device(sides = c("a", "b", "c", "d", "e", "f"),
#                       prob = c(1/6, 1/6, 1/6, 1/6, 1/6, 1/6))

# fruit_die <- device(sides = c("oranges", "plums", "apples",
#                     "bananas", "durians", "tomatoes"),
#                      prob = c(1/8, 1/8, 1/8, 1/8, 1/8, 3/8))

# loaded die
# loaded <- device(sides = c(1, 2, 3, 4, 5, 6), prob = c(.1, .1, .1, .1, .1, .5))

}
