% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{device}
\alias{device}
\title{Device}
\usage{
device(sides = c(1, 2), prob = c(0.5, 0.5))
}
\arguments{
\item{sides}{vector of sides}

\item{prob}{vector of prob}
}
\value{
an object of class \code{"device"}
}
\description{
creates an object of class \code{"device"}
}
\examples{
device1 <- device()
}
