% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{device}
\alias{device}
\alias{is.device}
\title{Device}
\usage{
device(sides = c(1, 2), prob = rep(0.5, 2))

is.device(x)
}
\arguments{
\item{sides}{vector of k ≥ 2 elements, by default numbers 1 and 2.}

\item{prob}{vector of probabilities for each side (all equal to 1/2 by default).}

\item{x}{an R object}
}
\value{
dev an object of class \code{"device"}
}
\description{
Creates an R object of class \code{"device"}
}
\examples{
# default
fair_coin <- device()

# another device
weird_die <- device(sides = c('i', 'ii', 'iii', 'iv'), prob = rep(1/4, 4))

#loaded device
loaded_die <- device(sides = 1:6, prob = c(0.075, 0.1, 0.125, 0.15, 0.20, 0.35))

}
\seealso{
\code{\link{roll}}
}
