% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{is.device}
\alias{is.device}
\title{is.device}
\usage{
is.device(x)
}
\arguments{
\item{x}{input object}
}
\value{
TRUE/FALSE if the input are objects or not
}
\description{
check if the class of input is device
}
