% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{is.device}
\alias{is.device}
\title{is.device}
\usage{
is.device(x)
}
\arguments{
\item{x}{object}
}
\value{
TRUE if it is of class "device", FALSE othewise
}
\description{
tests an argument if it is of class "device"
}
