% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{is.device}
\alias{is.device}
\title{Is.Device Function}
\usage{
is.device(x)
}
\arguments{
\item{x}{any object}
}
\value{
displays TRUE if the class of the object is device or FALSE is the class of the object is not device
}
\description{
checks if the item in the function is a device
}
\examples{
# fair_coin <- device()
# is.device(fair_coin)
# [1] TRUE

# loaded <- device(sides = c(1, 2, 3, 4, 5, 6), prob = c(.1, .1, .1, .1, .1, .5))
# is.device(loaded)
# [1] TRUE

}
