% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{is.device}
\alias{is.device}
\title{check device}
\usage{
is.device(x)
}
\arguments{
\item{fucntion}{of device}
}
\value{
TRUE or FALSE. If it is class of device TRUE, if not FALSE \code{"is.device"}
}
\description{
creates checking of an object of class \code{"is.device"}
}
\examples{
fair_coin <- device()
is.device(fair_coin)


}
