% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othermethods.R
\name{make_roll}
\alias{make_roll}
\title{Make Rolled Object}
\usage{
make_roll(device, flips)
}
\arguments{
\item{flips}{object of class flips}

\item{coin}{object of class device}
}
\description{
Constructor function for object "roll"
}
