% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-roll.R
\name{plot.rolls}
\alias{plot.rolls}
\title{Plot of object roll}
\usage{
\method{plot}{rolls}(x, ...)
}
\arguments{
\item{x}{an object of class}

\item{\dots}{arguments to be passed to/from other methods}
}
\description{
Plots the relative frequencies of a series of rolls
}
