% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{plot.rolls}
\alias{plot.rolls}
\title{Plot for rolls}
\usage{
\method{plot}{rolls}(rolled)
}
\arguments{
\item{rolled}{a "rolls" object}
}
\value{
a plot of sides and their frequencies
}
\description{
plots the frequencies of a rolls object
}
