% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.rolls}
\alias{plot.rolls}
\title{plot.rolls}
\usage{
\method{plot}{rolls}(roll, times)
}
\arguments{
\item{roll}{roll outcomes of input device}

\item{times}{total times of rolls}
}
\value{
side, count, probability summary of the input roll
}
\description{
plot input roll and times in bar chart
}
