% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-rolls.R
\name{plot.rolls}
\alias{plot.rolls}
\title{plot rolls}
\usage{
\method{plot}{rolls}(x, ...)
}
\arguments{
\item{a}{"rolls" object}
}
\value{
a barplot
}
\description{
produces a barplot of the result of a "rolls" object
}
