% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-roll.R
\name{plot.rolls}
\alias{plot.rolls}
\title{plot roll method}
\usage{
\method{plot}{rolls}(x, ...)
}
\arguments{
\item{x}{object of class "rolls"}
}
\value{
a barchart of proportions
}
\description{
takes 'rolls' objects and return a barchart of proportions
}
\examples{
 \dontrun{
 #roll the default die for 60 times
 device3 <- device()
 roll3 <- roll(device3, 60)

 #create barchard of proportions
 plot(roll3)}
}
