% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roller.R
\name{plot.rolls}
\alias{plot.rolls}
\title{Plot of object rolls}
\usage{
\method{plot}{rolls}(rolls, ...)
}
\arguments{
\item{rolls}{an R object of class \code{"rolls"}}

\item{\dots}{arguments to be passed to/from other methods}
}
\description{
In a barplot, plots the relative frequencies of the rolls against the sides of the device
}
\examples{

\dontrun{
# Create the basic default device object, a fair sided coin

device1 <- device()

# roll the basic device 100 times

roll100 <- roll(device = device1, times = 100)

plot(roll100)

# can add more to roll100

roll200 <- roll100 + 100

# can plot that too

plot(roll200)

}
}
