% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{plot.rolls}
\alias{plot.rolls}
\title{Plot of side frequencies}
\usage{
\method{plot}{rolls}(x)
}
\arguments{
\item{x}{a "rolls" object}
}
\description{
plots a bar graph of the side frequencies in a certain number of rolls of a device object
}
\examples{
#plotting the side frequencies in 7 rolls of a default device object
roll1 <- roll(device(), 7)
plot(roll1)

#plotting the side frequencies in 16 rolls of a non-default device object
roll2 <- roll(device(sides = c(1,2,3), prob = c(0.1,0.3,0.6)), 16)
plot(roll2)
}
