% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{plot.rolls}
\alias{plot.rolls}
\title{Plot Rolls}
\usage{
\method{plot}{rolls}(roll_d)
}
\arguments{
\item{roll_d}{an object of rolls}
}
\value{
a plot of sides and their frequencies
}
\description{
plots the relative frequencies of rolls object
}
