% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.rolls}
\alias{plot.rolls}
\title{Plot of object rolled}
\usage{
\method{plot}{rolls}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"roll"}}

\item{side}{number indicating which side of device to consider}
}
\description{
Plots the relative frequencies of a series of rolls
}
