% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-rolls.R
\name{plot.rolls}
\alias{plot.rolls}
\title{Plot of object roll}
\usage{
\method{plot}{rolls}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"roll"}}
}
\description{
Plots the relative frequencies of a series of roll
}
