% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{plot.rolls}
\alias{plot.rolls}
\title{Plot function}
\usage{
\method{plot}{rolls}(x, y = 50)
}
\arguments{
\item{x}{any input that is of the class "rolls"}

\item{y}{any input that should be the number of rolls the rolls object has (sample size)}
}
\value{
a plot of the proportions of the rolls object as the heights
}
\description{
takes in rolls object and outputs a plot of the proportions of the rolls object
}
\examples{
# string die
# str_die <- device(
#  sides = c('a', 'b', 'c', 'd', 'e', 'f'),
#  prob = c(0.075, 0.1, 0.125, 0.15, 0.20, 0.35))
# roll 20 times
# set.seed(123)
# str_rolls <- roll(str_die, times = 20)
# plot(str_rolls)
}
