% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{plot.rolls}
\alias{plot.rolls}
\title{barplot of rolling object}
\usage{
\method{plot}{rolls}(x, ...)
}
\arguments{
\item{rolling}{object}
}
\value{
barplot of relative frequencies and sides \code{"plot.rolls"}
}
\description{
using for getting
barplot of relative frequencies and sides \code{"plot.rolls"}
}
\examples{

}
