% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-roll.R
\name{plot.rolls}
\alias{plot.rolls}
\title{Plot of object roll}
\usage{
\method{plot}{rolls}(x, y)
}
\arguments{
\item{x}{an object of class \code{"device"}}

\item{y}{number of total rolls}
}
\description{
Plots the relative frequencies of a series of rolls
}
\examples{
 \dontrun{
 # create an object and roll it 50 times
 fair_die <- device(sides = 1:6, prob = rep(1/6, 6))
 fair_50rolls <- roll(fair_die, times = 50)

 plot(fair_50rolls)
 }
}
