% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{+.rolls}
\alias{+.rolls}
\title{Adding  rolls}
\usage{
\method{+}{rolls}(rolled, n)
}
\arguments{
\item{rolled}{a rolls object}

\item{n}{a number of rolls to add}
}
\value{
the updated rolls object
}
\description{
adds more rolls
}
