% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalmethods.R
\name{+.rolls}
\alias{+.rolls}
\title{+. rolls}
\usage{
\method{+}{rolls}(roll, incr)
}
\arguments{
\item{roll}{roll outcomes of input device}

\item{incr}{increase of roll times}
}
\value{
new roll information
}
\description{
addition method
}
