% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-methods-roll.R
\name{+.rolls}
\alias{+.rolls}
\title{roll addition method}
\usage{
\method{+}{rolls}(x, n)
}
\arguments{
\item{x}{object of class "rolls"}

\item{n}{n more times to be added to the rolls}
}
\value{
new 'rolls' object with added rolls
}
\description{
addition method to add more rolls
}
\examples{
device2 <- device()
roll2 <- roll(device2, 20)

#add 10 more rolls
roll2 + 10
}
