% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{+.rolls}
\alias{+.rolls}
\title{Addition Method for rolls}
\usage{
\method{+}{rolls}(x, incr)
}
\arguments{
\item{x}{a "rolls" object}

\item{incr}{the number of new rolls you want to add}
}
\description{
adds new rolls to a "rolls" object
}
\examples{
#example 1: We have a roll object with the default device (the sides are 1 and 2 with probabilities 0.5 and 0.5) and times = 7 (7 rolls) as the parameters.
roll1 <- roll(device(),7)
The output of roll1 + 10 will be a "rolls" object with 10 additional rolls.
So, if the output of roll1 is c(1,1,2,1,2,2,1), then the output of roll1 + 10 could be c(1,1,2,1,2,2,1,1,2,1,1,2,1,2,1,2,2).

#example 2: We have a roll object with device(sides = c(1,2,3), prob = c(0.1,0.3,0.6)) and times = 16 as the parameters.
roll2 <- roll(device(sides = c(1,2,3), prob = c(0.1,0.3,0.6)),times = 16)
The output of roll2 + 5 will be a "rolls" object with 5 additional rolls.
So, if the output of roll2 is c(1,2,1,3,2,3,2,1,1,3,2,2,3,1,2,3), then the output of roll1 + 10 could be c(1,2,1,3,2,3,2,1,1,3,2,2,3,1,2,3,3,1,2,3,3).
}
