% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-methods-roll.R
\name{+.rolls}
\alias{+.rolls}
\title{add method}
\usage{
\method{+}{rolls}(x, y)
}
\arguments{
\item{x}{list of class rolls}

\item{y}{integer for additional number of rolls}
}
\value{
adds more rolls
}
\description{
adds certain number of rolls according to
user input
}
