% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{+.rolls}
\alias{+.rolls}
\title{Add rolls}
\usage{
\method{+}{rolls}(roll2, n)
}
\arguments{
\item{roll2}{an object of type rolls}

\item{n}{how many more rolls to add}
}
\value{
the new rolls object with updates
}
\description{
addition of more rolls
}
