% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{+.rolls}
\alias{+.rolls}
\title{`+.rolls` Addition Method/Function}
\usage{
\method{+}{rolls}(obj, incr)
}
\arguments{
\item{obj}{rolls object}

\item{incr}{the number of more rolls to the rolls object}
}
\value{
a new rolls object with a greater sample size
}
\description{
adds a given number of more rolls to a rolls object
}
\examples{
#' # string die
# str_die <- device(
#  sides = c('a', 'b', 'c', 'd', 'e', 'f'),
#  prob = c(0.075, 0.1, 0.125, 0.15, 0.20, 0.35))
# roll 20 times
# set.seed(123)
# str_rolls <- roll(str_die, times = 20)
# newstr_rolls <- str_rolls + 10
}
