% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{print.device}
\alias{print.device}
\title{device print}
\usage{
\method{print}{device}(x)
}
\arguments{
\item{device}{object}
}
\value{
none
}
\description{
displays the class of the object as well as a tabular display of the sides and probabilities
}
