% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{print.device}
\alias{print.device}
\title{print.device}
\usage{
\method{print}{device}(x, ...)
}
\value{
print sides and probability of device
}
\description{
print device information in an organized manner
}
