% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{print.rolls}
\alias{print.rolls}
\title{rolls}
\usage{
\method{print}{rolls}(list)
}
\arguments{
\item{list}{input roll list}
}
\value{
print rolls info in an organized manner
}
\description{
print rollss information in an organized manner
}
