% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{print.rolls}
\alias{print.rolls}
\title{rolls print}
\usage{
\method{print}{rolls}(x)
}
\arguments{
\item{rolls}{object}
}
\value{
none
}
\description{
displays the class of the object as well as the generated rolls
}
