% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{print.summary.rolls}
\alias{print.summary.rolls}
\title{print rolls summary}
\usage{
\method{print}{summary.rolls}(x, ...)
}
\arguments{
\item{a}{summary of rolls}
}
\value{
none
}
\description{
prints the summary of a number of rolls
}
