% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\alias{is.roll}
\alias{as.roll}
\title{roll}
\usage{
roll(x, times = 1)

is.roll(x)

as.roll(x)
}
\arguments{
\item{x}{device object (a vector)}

\item{times}{number of rolls}

\item{prob}{vector of probabilities for each side of the device}

\item{x}{an R object}
}
\value{
vector of rolls
#' @export
}
\description{
Simulates rolling a device a given number of times
}
