% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{roll}
\alias{roll}
\title{Roll a device}
\usage{
roll(device, times = 1)
}
\arguments{
\item{device}{object of class \code{"device"}}

\item{times}{number of rolls}
}
\value{
an object of class \code{"roll"} with the following elements:

\item{rolls}{vector of rolls}

\item{sides}{vector of device \code{"sides"}}

\item{prob}{vector of device \code{"prob"}}

\item{total}{total number of rolls}
}
\description{
Creates an object of class \code{"roll"}
}
\examples{
fair_die <- device()

# roll a fair dice:
roll(fair_die, times = 50)

}
