% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{roll}
\usage{
roll(dev, times = 1)
}
\arguments{
\item{sides}{sides info of the object}

\item{prob}{prob info of the object}
}
\value{
device information, including sides and probabilities
}
\description{
return information of the roll
}
