% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{roll}
\alias{roll}
\title{roll constructor}
\usage{
roll(device, times = 1)
}
\arguments{
\item{device}{(object of class 'device')}

\item{number}{of times rolled (int)}
}
\value{
a "rolls" object
}
\description{
creates a roll object
}
