% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{roll}
\usage{
roll(x, times = 1)
}
\arguments{
\item{x}{object of class "device"}

\item{times}{number of times to roll the device (default = 1)}
}
\value{
a list of rolls, sides, prob, total
}
\description{
takes a device and a number of times and return rolls
}
\examples{
device1 <- device()

#roll the device 20 times
roll20 <- roll(device1, 20)
}
