% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roller.R
\name{roll}
\alias{roll}
\alias{is.rolls}
\title{Roll}
\usage{
roll(device, times = 1)

is.rolls(src)
}
\arguments{
\item{device}{an R object of class \code{"device"}}

\item{times}{the number of times the user wishes to roll the device}

\item{src}{any R object}
}
\value{
an object of class \code{"rolls"}

an item \code{"rolls"}: a vector of length(times) containing the rolls of the device

an item \code{"sides"}: a vector of the names or numbers of the sides of the rolled device

an item \code{"prob"}: a vector of probabilites for each side that the rolled device has

an item \code{"total"}: a single number equal to the number of times the device was rolled, which is equal to parameter times

boolean value indicating if the source object is of class \code{"rolls"}
}
\description{
This function creates the rolls that the device contains, and rolls them as many times as the user defines
}
\examples{

# creating the device that makes the roll
device1 <- device()

# Rolls the device 50 times
roll50 <- roll(device, 50)

# the user can add more rolls to the basic device:
roll150 <- roll50 + 100

}
