% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{roll}
\alias{roll}
\title{Roll}
\usage{
roll(device, times)
}
\arguments{
\item{device}{a device object}

\item{times}{the number of times device is rolled}
}
\value{
a list containing vectors with the outputs of the rolls, the sides of the "device" object, and the probabilities of the sides, as well as the total number of rolls
}
\description{
returns the result of "rolling" a device object a certain number of times
}
\examples{
#example 1: default device and 7 rolls
roll1 <- roll(device(), 7)


#example 2: non-default device and 16 rolls
roll2 <- roll(device(sides = c(1,2,3), prob = c(0.1,0.3,0.6)), 16)
}
