% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{Roll Function}
\usage{
roll(device, times = 1)
}
\arguments{
\item{device}{object of class device}

\item{times}{integer}
}
\value{
prints the class name and the number result of rolls
}
\description{
Rolls an object. By default, the number of rolls is 1
}
\examples{

#create a fair coin
fair_coin = device(c("heads", "tails"), c(0.5, 0.5))

#use roll to simulate a coin toss
roll(fair_coin)

#toss same coin 15 times
roll(fair_coin, times = 15)
}
