% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{roll}
\alias{roll}
\title{Roll}
\usage{
roll(device, times = 1)
}
\arguments{
\item{device}{an object of class "device"}

\item{times}{number of rolls}
}
\value{
an object of class "roll"

\item{rolls}{vector of rolls}

\item{sides}{vector of device \code{"sides"}}

\item{prob}{vector of device \code{"prob"}}

\item{total}{total number of rolls}
}
\description{
Create an object of class "roll"
}
