% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R, R/summary.rolls.R
\name{roll}
\alias{roll}
\alias{is.roll}
\alias{summary.roll}
\title{Roll a device}
\usage{
roll(device, times = 1)

is.roll(x)

\method{summary}{roll}(x, ...)
}
\arguments{
\item{device}{object of class "device"}

\item{times}{number of times to roll the device (default value of 1)}

\item{x}{an R object}

\item{x}{an object "roll"}

\item{further}{arguments ignored}
}
\value{
an object of class "rolls" with the following elements:

rolls vector of rolls

sides vector of device "sides"

prob vector of device "prob"

total total number of rolls(i.e. times)
}
\description{
Creates an object of class "rolls"
}
