% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{roll}
\alias{roll}
\title{roll a device}
\usage{
roll(device, times = 1)
}
\arguments{
\item{device}{object of class "device"}

\item{times}{number of times to roll the device (default value of 1)}
}
\value{
an object of class \code{"roll"} with the following elements:

\item{rolls}{vector of rolls}

\item{sides}{vector of device \code{"sides"}}

\item{prob}{vector of device \code{"prob"}}

\item{total}{total number of rolls}
}
\description{
Creates an object of class \code{"roll"}
}
