% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{roll}
\alias{roll}
\title{Rolls function}
\usage{
roll(obj, times = 1)
}
\arguments{
\item{obj}{object of class \code{"roll"}}

\item{times}{number of tosses}
}
\value{
an object of class \code{"roll"} with the following elements:

\item{rolls}{vector of outputs of the rolls}

\item{sides}{vector of "device" object \code{"sides"}}

\item{prob}{vector of "device" object \code{"prob"}}

\item{total}{total number of rolls ie times}
}
\description{
creates an object of class \code{"roll"}
}
\examples{

# default roll
fair_coin <- device()
default_roll <- roll(fair_coin)

# using a regular die rolling 50 times
regular_die <- device(sides = c("a", "b", "c", "d", "e", "f"), prob = c(1/6, 1/6, 1/6, 1/6, 1/6, 1/6))
regular50rolls <- roll(obj = regular_die, times = 50)

# using a weird die 100 times
weird_die <- device(sides = c("oranges", "plums", "apples", "bananas", "durians", "tomatoes"), prob = c(1/8, 1/8, 1/8, 1/8, 1/8, 3/8))
weird100rolls <- roll(obj = weird_die, times = 100)
}
