% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{Roll a device}
\usage{
roll(device, times = 1)
}
\arguments{
\item{device}{object of class \code{"device"}}

\item{times}{number of rolls}
}
\value{
output, an object of class \code{"rolls"}
}
\description{
Creates an object of class \code{"rolls"}
}
\examples{
fair_die <- device()

# roll a fair die 50 times
fair50 <- roll(fair_die, times = 50)

# add 5 more rolls
fair55 <- fair50 + 5
}
