% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{[.rolls}
\alias{[.rolls}
\title{Extraction for rolls}
\usage{
\method{[}{rolls}(rolled, n)
}
\arguments{
\item{rolled}{a rolls objectsumm}

\item{n}{an index}
}
\value{
the item at nth index
}
\description{
extracts objects from a "rolls" object
}
