% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalmethods.R
\name{[.rolls}
\alias{[.rolls}
\title{[.rolls}
\usage{
\method{[}{rolls}(roll, i)
}
\arguments{
\item{roll}{roll outcomes of input device}

\item{i}{index of numbers being extracted}
}
\value{
extracted number
}
\description{
extract elements
}
