% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-roll-methods.R
\name{[.rolls}
\alias{[.rolls}
\title{rolls extraction}
\usage{
\method{[}{rolls}(x, i)
}
\arguments{
\item{a}{rolls object}
}
\value{
the value of a given roll
}
\description{
extracts the value of a given roll
}
