% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-methods-roll.R
\name{[.rolls}
\alias{[.rolls}
\title{roll extraction method}
\usage{
\method{[}{rolls}(x, i)
}
\arguments{
\item{x}{object of class "rolls"}

\item{i}{index of the rolls to be changed}
}
\value{
value of the rolls at the given index
}
\description{
extract the value of given index in the rolls
}
\examples{
device2 <- device()
roll2 <- roll(device2, 20)

#get the 10th result of roll2
roll2[10]
}
