% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{[.rolls}
\alias{[.rolls}
\title{Extraction Method for rolls}
\usage{
\method{[}{rolls}(x, rolls_index)
}
\arguments{
\item{x}{a "rolls" object}

\item{rolls_index}{an index "i" corresponding to the ith roll of the device object}
}
\description{
extracts the side of a "device" object that was rolled on a specific roll corresponding to the inputted index
}
\examples{
#example 1: We have a roll object with the default device (the sides are 1 and 2 with probabilities 0.5 and 0.5) and times = 7 (7 rolls) as the parameters.
roll1 <- roll(device(),7)
The output of roll1[4] is the side rolled on the fourth roll.

#example 2: We have a roll object with device(sides = c(1,2,3), prob = c(0.1,0.3,0.6)) and times = 16 as the parameters.
roll2 <- roll(device(sides = c(1,2,3), prob = c(0.1,0.3,0.6)),times = 16)
The output of roll2[12] is the side rolled on the twelfth roll.
}
