% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-methods-roll.R
\name{[.rolls}
\alias{[.rolls}
\title{Extraction method}
\usage{
\method{[}{rolls}(x, i)
}
\arguments{
\item{x}{list of class rolls}

\item{i}{integer}
}
\value{
value of a particular roll
}
\description{
extract value of given roll
}
