% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{[.rolls}
\alias{[.rolls}
\title{Extract Method}
\usage{
\method{[}{rolls}(roll2, n)
}
\arguments{
\item{roll2}{an object of rolls}

\item{n}{an index}
}
\value{
the value at the ith index
}
\description{
extracts specified values from a roll
}
