% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{[.rolls}
\alias{[.rolls}
\title{`[.rolls` Extraction Method/Function}
\usage{
\method{[}{rolls}(x, i)
}
\arguments{
\item{x}{rolls object}

\item{i}{the location/index as a number}
}
\value{
the numeric/character element of that roll at the specific index
}
\description{
extracts the specified value based on location/index from rolls object
}
\examples{
#' # string die
# str_die <- device(
#  sides = c('a', 'b', 'c', 'd', 'e', 'f'),
#  prob = c(0.075, 0.1, 0.125, 0.15, 0.20, 0.35))
# roll 20 times
# set.seed(123)
# str_rolls <- roll(str_die, times = 20)
# str_rolls[3]
# str_rolls[4]
}
