% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{[<-.rolls}
\alias{[<-.rolls}
\title{Replacement for rolls}
\usage{
\method{[}{rolls}(rolled, index) <- value
}
\arguments{
\item{rolled}{a rolls object}

\item{index}{an index of roll to replace}

\item{value}{an item to replace it with}
}
\value{
the updated rolled object
}
\description{
replaces a value of a given roll
}
