% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalmethods.R
\name{[<-.rolls}
\alias{[<-.rolls}
\title{[<- rolls}
\usage{
\method{[}{rolls}(x, i) <- value
}
\arguments{
\item{x}{roll outcomes of input device}

\item{i}{index of element being changed}

\item{value}{value of element in replacement}
}
\value{
new roll information
}
\description{
replacement method
}
