% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-roll-methods.R
\name{[<-.rolls}
\alias{[<-.rolls}
\title{rolls replacement}
\usage{
\method{[}{rolls}(x, i) <- value
}
\arguments{
\item{a}{rolls object}
}
\value{
none
}
\description{
replaces the value of a given roll
}
