% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-methods-roll.R
\name{[<-.rolls}
\alias{[<-.rolls}
\title{roll replacement method}
\usage{
\method{[}{rolls}(x, i) <- value
}
\arguments{
\item{x}{object of class "rolls"}

\item{i}{index of the rolls to be changed}

\item{value}{new value to be inputted}
}
\value{
new rolls with replaced value
}
\description{
replacement method to replace the value of a given roll
}
\examples{
device2 <- device()
roll2 <- roll(device2, 20)

#replace 3rd result to 2
roll2[2] <- 2
}
