% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-methods-roll.R
\name{[<-.rolls}
\alias{[<-.rolls}
\title{Replacement method}
\usage{
\method{[}{rolls}(x, i) <- value
}
\arguments{
\item{x}{list of class rolls}

\item{i}{integer}

\item{value}{new value of}
}
\value{
replaces the value of vector, at a particular index
}
\description{
replace value of given roll with another value
}
