% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{[<-.rolls}
\alias{[<-.rolls}
\title{Replace Method}
\usage{
\method{[}{rolls}(roll2, index) <- value
}
\arguments{
\item{roll2}{an object of rolls}

\item{index}{the index of the value to replace}

\item{value}{what will take its place}
}
\value{
the new rolled object
}
\description{
replaces specified values from a roll
}
