% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{[<-.rolls}
\alias{[<-.rolls}
\title{`[<-.rolls` Replacement Method/Function}
\usage{
\method{[}{rolls}(x, i) <- value
}
\arguments{
\item{x}{rolls object}

\item{i}{the location/index as a number}

\item{value}{the numeric/character element that will replace the ith character of the rolls object}
}
\value{
the numeric/character element of that roll at the specific index replaced with the value
}
\description{
extracts the specified value based on location/index from rolls object and replaces the value with one of the sides of the roll object
}
\examples{
#' # string die
# str_die <- device(
#  sides = c('a', 'b', 'c', 'd', 'e', 'f'),
#  prob = c(0.075, 0.1, 0.125, 0.15, 0.20, 0.35))
# roll 20 times
# set.seed(123)
# str_rolls <- roll(str_die, times = 20)
# str_rolls[3] <- "a"
# str_rolls[3]
# str_rolls[4] <- "e"
# str_rolls[4]
}
