% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{summary.rolls}
\alias{summary.rolls}
\title{Summary for rolls (summary.rolls)}
\usage{
\method{summary}{rolls}(rolled)
}
\arguments{
\item{rolled}{a rolls object}
}
\value{
a summary.rolls object containing data frame of side, count, and prop
}
\description{
returns an object "summary.rolls" which is a list of a data frame
describing the sides, count and proportion or the rolled device
}
