% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rolls.R
\name{summary.rolls}
\alias{summary.rolls}
\title{summary}
\usage{
\method{summary}{rolls}(roll)
}
\arguments{
\item{roll}{roll outcomes of input device}
}
\value{
side, count, probability summary of the input roll
}
\description{
return summary of the roll
}
