% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{summary.rolls}
\alias{summary.rolls}
\title{rolls summary}
\usage{
\method{summary}{rolls}(x, ...)
}
\arguments{
\item{a}{rolls object}
}
\value{
a table summarizing the rolls
}
\description{
summarizes the results of a number of rolls
}
