% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-roll.R
\name{summary.rolls}
\alias{summary.rolls}
\title{summary roll method}
\usage{
\method{summary}{rolls}(x, ...)
}
\arguments{
\item{x}{object of class "rolls"}
}
\value{
a list containing a data frame with side, count, and prop
}
\description{
takes 'rolls' objects and return object 'summary rolls'
}
\examples{
device2 <- device()
roll2 <- roll(device2, 20)

#obtain summary of roll
summary(roll2)
}
