% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{summary.rolls}
\alias{summary.rolls}
\title{Summary method for class Roll}
\usage{
\method{summary}{rolls}(x)
}
\arguments{
\item{x}{of class 'rolls'}
}
\value{
object(list) with summary for rolls()
}
\description{
Creates Summary for Roll Function
}
