% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{summary.rolls}
\alias{summary.rolls}
\title{summary.rolls}
\usage{
\method{summary}{rolls}(rolled)
}
\arguments{
\item{rolled}{an object of rolls}
}
\value{
a dataframe with the summary of rolls containing
four elements
}
\description{
returns the content of rolls in a dataframe
}
