% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{summary.rolls}
\alias{summary.rolls}
\title{Summary function}
\usage{
\method{summary}{rolls}(x)
}
\arguments{
\item{x}{any input that is of the class "rolls"}
}
\value{
object class name "summary.rolls" and data frame with the "sides", "counts", and "prop" of the rolls
}
\description{
takes in rolls object and outputs the object class name "summary.rolls" and data frame with the "sides", "counts", and "prop" of the rolls
}
\examples{
# roll fair 8-sided die
# set.seed(123)
# fair_dev <- device(sides = letters[1:8], prob = rep(1/8, 8))
# fair500 <- roll(fair_dev, times = 500)
# summary method
# summary(fair500)
## summary "rolls"
##
## side count prop
## 1 a 63 0.126
## 2 b 54 0.108
## 3 c 73 0.146
## 4 d 69 0.138
## 5 e 69 0.138
## 6 f 51 0.102
## 7 g 65 0.130
## 8 h 56 0.112
# Another example
# roll 50 times
# set.seed(123)
# roll fair die 50 times
# fair_die <- device(sides = 1:6, prob = rep(1/6, 6))
# set.seed(123)
# fair_50rolls <- roll(fair_die, times = 50)
## summary(fair_50rolls)
##   side count prop
## 1    1    11 0.22
## 2    2     8 0.16
## 3    3     9 0.18
## 4    4     8 0.16
## 5    5     7 0.14
## 6    6     7 0.14
}
