% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{summary.rolls}
\alias{summary.rolls}
\title{summary of rolling object}
\usage{
\method{summary}{rolls}(x, ...)
}
\arguments{
\item{rolling}{object}
}
\value{
summary of count, prop, and sides \code{"summary.rolls"}
}
\description{
summary of rolling object and get data frame of sides, prob,
rolls, and total \code{"summary.rolls"}
}
\examples{


}
